* manage control elements in CCCH PDU through rx_sdu function * add event handler for remote (online) configuration and status monitoring * update openair config generator (OCG) and scneario descriptor (OSD) with new features contributed by A. Gomes (one source) * fix the issue with netlink when benchmarking oaisim preformance * pre-ci passed except for eMBMS
git-svn-id: http://svn.eurecom.fr/openair4G/trunk@5299 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ openair2/COMMON/platform_types.h
@@ -36,7 +36,7 @@ typedef uint32_t              frame_t;
 typedef int32_t               sframe_t;
 typedef uint32_t              sub_frame_t;
 typedef uint8_t               module_id_t;
-typedef int8_t                smodule_id_t;
+typedef int16_t                smodule_id_t;
 typedef uint16_t              rb_id_t;
 typedef uint16_t              srb_id_t;
